trigger_agent/requirements.txt
fastapi
uvicorn[kcp]
kafka-python
pydantic
sqlalchemy
psycopg2-binary